test: readme revisions and docstring updates [TESTENG-5]
