+++ webui/react/src/e2e/README.md
@@ -1,37 +1,49 @@
 # E2E
 
+Our goals are to keep test cases realiable, stable, maintainable, and readable.
+
 ## Framework
 
 Deteremined AI uses [Playwright 🎭](https://playwright.dev/).
 
-## How to run locally
+## Setup
 
-- Create `.env` file in `webui/react` like `webui/react/.env`
-- Add env variables `PW_USER_NAME`, `PW_PASSWORD`, and `PW_SERVER_ADDRESS` (`PW_` prefix stands for Playwright)
-  - `PW_USER_NAME`: user name for determined account
-  - `PW_PASSWORD`: password for determined account
-  - `PW_SERVER_ADDRESS`: API server address
-  - `PW_DET_PATH`: path to `det` if not already in path
-  - `PW_DET_MASTER`: typically http://localhost:8080"
-- Run `npx playwright install`
-- Run `SERVER_ADDRESS={set server address} npm run build` in `webui/react`
-  - It is `SERVER_ADDRESS` here. not `PW_SERVER_ADDRESS`, but the both values should be the same
-- Run `npm run e2e` in `webui/react`. Use `-- --project=<browsername>` to run a specific browser.
+Everything you need before running tests
 
-\*\*Avoid using `make` command because it does not take env variables
+### `.env`
 
-### Quick start testing using det deploy
+Create `.env` file in `webui/react` like `webui/react/.env` and env variables. (`PW_` prefix stands for Playwright)
 
-If you don't want to use dev cluster, you can use det deploy to initiate the backend. These commands should run and pass tests on chrome:
+- `PW_USER_NAME`: user name for determined account
+- `PW_PASSWORD`: password for determined account
+- `PW_SERVER_ADDRESS`: API server address
+- `PW_DET_PATH`: path to `det` if not already in path
+- `PW_DET_MASTER`: typically <http://localhost:8080>", used for CLI commands
 
-1. `det deploy local cluster-up --det-version="0.29.0" --no-gpu --master-port=8080`
-   - Use whatever det-version you want here.
+### Playwright
+
+Run `npx playwright install`
+
+### Determined
+
+Pick between live and static
+
+#### Live Changes
+
+1. `npm run start` `--prefix webui/react`
+2. `conda activate det &&` `devcluster`
+
+#### Static Build and Cluster Up
+
+1. `det deploy local cluster-up --no-gpu --master-port=8080`
 2. `SERVER_ADDRESS="http://localhost:3001" npm run build --prefix webui/react`
-3. Optional if you want an experiment created for the test: `det experiment create ./examples/tutorials/mnist_pytorch/const.yaml ./examples/tutorials/mnist_pytorch/`
-4. Optional `npm run preview --prefix webui/react` to run the preview app. Won't be used if `CI=true`.
-   1. Consider running `npm run start --prefix webui/react -- --port=3001` for live changes if you're editing page models. The other command will constantly throw build errors if you're editing tests and test hooks at the same time. We use port `3001` because that's the port playwright is configured to use.
-5. To run the tests: `PW_SERVER_ADDRESS="http://localhost:3001"  PW_USER_NAME="admin" PW_PASSWORD="" npm run e2e --prefix webui/react`
-   - Provice `-- -p=firefox` to choose one browser to run on. Full list of projects located in [playwright config](/webui/react/playwright.config.ts).
+3. Optional `npm run preview --prefix webui/react` to run the preview app. Won't be used if `CI=true`.
+4. To run the tests: `PW_SERVER_ADDRESS="http://localhost:3001"  PW_USER_NAME="admin" PW_PASSWORD="" npm run e2e --prefix webui/react`
+   - Provide `-- -p=firefox` to choose one browser to run on. Full list of projects located in [playwright config](/webui/react/playwright.config.ts).
+
+## Run Tests
+
+Use the [VS Code extension](https://marketplace.visualstudio.com/items?itemName=ms-playwright.playwright) to run individual tests. You can also use the commandline, here are [the docs](https://playwright.dev/docs/running-tests).
 
 ## CI
 
@@ -41,9 +53,13 @@ We use `mcr.microsoft.com/playwright` for [docker container](https://playwright.
 Update the docker image version along with Playwright version.
 
 - PW_DET_PATH=/tmp/venv/bin/det"
-- PW_SERVER_ADDRESS=http://localhost:3001"
+- PW_SERVER_ADDRESS=<http://localhost:3001>"
 - PW_USER_NAME=admin
 - PW_PASSWORD=
-- PW_DET_MASTER=http://localhost:8082"
-- DET_WEBPACK_PROXY_URL=http://localhost:8082"
+- PW_DET_MASTER=<http://localhost:8082>"
+- DET_WEBPACK_PROXY_URL=<http://localhost:8082>"
 - DET_WEBSOCKET_PROXY_URL=ws://localhost:8082"
+
+## Appendix
+
+- [Page model Readme](./models/README.md)
